Changed the while loop to function for specific number of iterations.
Doing this helps in getting the answers quickly and are very close to actual answers, and avoids getting stuck in a long loop to find global minima. Also, I have changed the cost function to print the values of b, m, and the average cost after each iteration. This helps in getting to know what is happening behind the scenes.